didnt properly add tabs

+++ doubly_linked_lists/0-print_dlistint.c
@@ -2,13 +2,13 @@
 
 /**
  * prints all elements of dlistint_t
- * @h = ptr to head
+ * @h = head dlistint_t
  * Return: number of nodes
  */
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t c = 0;
+	unsigned int c = 0;
 
 	while (h != NULL)
 	{
